Rewrite/refactor most of the logic to be better (v2)
Notable outward change is that the invocation is different + returns a nested Dry::Struct rather than the old series of separate method calls and hashes. We also return a little more information about the statement because… why shouldn’t we?

Internally things are a little different + have been tidied up in an attempt to make it way less brittle than it used to be.